[Feat] 소셜 로그인을 위한 OIDC 도입

+++ back/src/main/java/com/back/auth/adapter/in/web/dto/AuthMemberResponse.java
@@ -4,8 +4,9 @@ import com.back.member.domain.Member;
 
 /** 인증 API에서 사용하는 회원 정보 응답 DTO. */
 public record AuthMemberResponse(
-    Integer id, String email, String nickname, String role, String status) {
+    Long id, String email, String nickname, String role, String status) {
 
+  /** Member 엔티티를 인증 응답용 DTO로 변환한다. */
   public static AuthMemberResponse from(Member member) {
     return new AuthMemberResponse(
         member.getId(),
